code/requirements.txt
docker==3.7.2
Flask==1.0.0
gunicorn
nuvla-api